+++ TT/TT.h
@@ -6,25 +6,25 @@
 #include <vector>
 
 struct TTEntry {
-    int value;                   // Evaluation score
+    float value;                 // Evaluation score
     int depth;                   // Depth of the search when this position was evaluated
     int nodeType;                // Type of node (upperbound, exact, lowerbound) (1, 2, 3) respectively
     chess::Move bestMove;        // Best move found from this position
 
     TTEntry() 
-        : value(0), depth(0), nodeType(1), bestMove(chess::Move::NO_MOVE) {}  // Default constructor
+        : value(0.0f), depth(0), nodeType(1), bestMove(chess::Move::NO_MOVE) {}  // Default constructor
 
     TTEntry(int v, int d, int nt, chess::Move bm)
         : value(v), depth(d), nodeType(nt), bestMove(bm) {}
 };
 
 struct KM {
-    int value;                      // Evaluation score
+    float value;                    // Evaluation score
     int depth;                      // Depth of the search when this position was evaluated
     chess::Move KillerMove;         // Best move found from this position
 
     KM() 
-        : value(0), depth(0), KillerMove(chess::Move::NO_MOVE) {} // Default constructor
+        : value(0.0f), depth(0), KillerMove(chess::Move::NO_MOVE) {} // Default constructor
 
     KM(int v, int d, chess::Move bm)
         : value(v), depth(d), KillerMove(bm) {}
